+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/messenger/compose/view/roomlist/room/RoomListElement.kt
@@ -3,8 +3,13 @@ package de.connect2x.messenger.compose.view.roomlist.room
 import androidx.compose.animation.core.Spring
 import androidx.compose.animation.core.VisibilityThreshold
 import androidx.compose.animation.core.spring
+import androidx.compose.foundation.LocalIndication
 import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
+import androidx.compose.foundation.gestures.detectTapGestures
+import androidx.compose.foundation.hoverable
+import androidx.compose.foundation.indication
+import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.foundation.interaction.PressInteraction
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
@@ -16,8 +21,10 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.unit.IntOffset
 import androidx.compose.ui.unit.dp
 import de.connect2x.messenger.compose.view.DI
@@ -56,8 +63,7 @@ class RoomListElementContainerViewImpl : RoomListElementContainerView {
         val selectedRoomId = roomListViewModel.selectedRoomId.collectAsState().value
         val roomName = roomListElementViewModel.roomName.collectAsState().value
         val isInvite = roomListElementViewModel.isInvite.collectAsState().value
-        Modifier
-            .fillMaxWidth()
+        val interactionSource = remember { MutableInteractionSource() }
         Box(
             Modifier.animateItem(
                 fadeInSpec = null,
@@ -67,8 +73,38 @@ class RoomListElementContainerViewImpl : RoomListElementContainerView {
                     visibilityThreshold = IntOffset.VisibilityThreshold
                 )
             )
-                .clickable(enabled = roomName != null && (isInvite == null || isInvite == false)) {
-                    roomListViewModel.selectRoom(roomId)
+                // TODO: Replace if Modifier.clickable works again
+                // For explanation:
+                // For some reason Modifier.clickable leads to the LazyColumn not recomposing the item where the click
+                // happened:
+                // LazyColumn {
+                //     items(100) {
+                //         println("rerun: $it")
+                //         Text(
+                //             text = "$it",
+                //             modifier = Modifier.clickable {
+                //                 println("clicked: $it")
+                //             }
+                //         )
+                //      }
+                //  }
+                .indication(interactionSource, LocalIndication.current)
+                .hoverable(interactionSource, enabled = roomName != null && isInvite != null && isInvite == false)
+                .pointerInput(roomName, roomId, isInvite) {
+                    if (roomName == null || isInvite == null || isInvite == true) return@pointerInput
+
+                    detectTapGestures(
+                        onPress = { offset ->
+                            val press = PressInteraction.Press(offset)
+                            interactionSource.emit(press)
+                            val released = tryAwaitRelease()
+                            interactionSource.emit(PressInteraction.Release(press))
+
+                            if (released) {
+                                roomListViewModel.selectRoom(roomId)
+                            }
+                        }
+                    )
                 }
                 .background(if (roomId == selectedRoomId) { MaterialTheme.components.roomListSelection.color } else Color.Unspecified)
                 .buttonPointerModifier(enabled = isInvite == null || isInvite == false)